pyproject.toml
[project]
name = "judge-me-now"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.67.0",
    "gtts>=2.5.4",
    "pydantic>=2.12.5",
    "streamlit>=1.55.0",
]
